Add support to Composer `archive:zip` script (#54)
This update adds support to the new `archive:zip` script that will be
used to create archive of the plugin. This command will replace the `npm
run plugin-zip` command from `wp-scripts` which rely on the npm pack and
has issues that we can't change. For example, it cannot exclude
the `README.md` file.